requirements.txt
py-cord==2.0.0b1
xmltodict==0.12.0
fandom-py==0.2.1